Add files via upload
capitalize():

This method capitalizes the first character of a string and converts the rest of the characters to lowercase. For instance, if the original string is "hello, and welcome to my world.", after applying capitalize(), it becomes "Hello, and welcome to my world."
casefold():

The casefold() method returns a string where all characters are converted to lowercase. This method is similar to lower(), but it handles more cases, especially for non-ASCII characters.
center(width[, fillchar]):

center() aligns the string centrally, padding the sides with a specified character (space by default) until the total width is reached.
count(sub[, start[, end]]):

The count() method returns the number of occurrences of a substring within the string. It allows specifying optional start and end positions to limit the search.
encode(encoding="utf-8", errors="strict"):

encode() returns the encoded version of the string using the specified encoding. It's useful for converting strings to bytes.
endswith(suffix[, start[, end]]):

endswith() checks if a string ends with a specified suffix. It returns True if the string ends with the given suffix, otherwise False.
expandtabs(tabsize=8):

The expandtabs() method replaces tab characters ('\t') with spaces. By default, it uses a tab size of 8 spaces, but you can specify a different size.
find(sub[, start[, end]]):

find() searches for a substring within the string and returns the lowest index of its occurrence. If the substring is not found, it returns -1.
**format(*args, kwargs):

format() formats the string by replacing placeholders (marked by curly braces) with corresponding arguments or keyword arguments passed to the method.
index(sub[, start[, end]]):

Similar to find(), index() finds the index of the first occurrence of a substring in the string. However, if the substring is not found, it raises a ValueError.
These methods provide various functionalities for manipulating strings in Python.